Deleted cat.cpp, moved the expression builder into SP folder
