--- source/rococo/sexy.domme.test/cat.cpp
@@ -1,55 +1,0 @@
-#include <Domme/Animals/ICat.hpp>
-
-#include <rococo.domme.h>
-
-using namespace Rococo;
-using namespace Rococo::Domme;
-
-namespace Rococo::Animals::Implementation
-{
-	struct Cat : ICatSupervisor
-	{
-		DommeObject D;
-
-		int makeBiscuitsIndex = -1;
-
-		Cat(ScriptingResources& _scripting, cstr sourceName) : D(_scripting, sourceName, "Rococo.Animals", "ICat")
-		{
-			makeBiscuitsIndex = D.GetMethodIndex("MakeBiscuits", 2, 0);
-		}
-
-		~Cat()
-		{
-		}
-
-		void MakeBiscuits(int nBiscuits, float dt) override
-		{
-			REGISTER_DOMME_CALL(D);
-
-			D.Push(nBiscuits);
-			D.Push(dt);
-			D.CallVirtualMethod(makeBiscuitsIndex);
-			D.PopBytes(sizeof nBiscuits + sizeof dt);
-
-			VALIDATE_REGISTERS;
-		}
-
-		void _Terminate() override
-		{
-			D.Terminate();
-		}
-
-		void _Free() override
-		{
-			delete this;
-		}
-	};
-}
-
-namespace Rococo::Animals
-{
-	ICatSupervisor* CreateCat(ScriptingResources& scripting, cstr sourceFile)
-	{
-		return new Rococo::Animals::Implementation::Cat(scripting, sourceFile);
-	}
-}
