+++ source/rococo/sexy/SS/sexy.script/sexy.global.functions.cpp
@@ -775,7 +775,7 @@ namespace Rococo
 		{
 		}
 
-		IAllocator& Allocator() override
+		IAllocator& Allocator() const override
 		{
 			return allocator;
 		}
